+++ sw.js
@@ -1,4 +1,4 @@
-const CACHE_NAME = '24points-v21';
+const CACHE_NAME = '24points-v22';
 const urlsToCache = ['.', 'index.html', 'main.js', 'manifest.json'];
 
 self.addEventListener('install', e => {